fix: fix hrmp script and share proposal construction logic
