+++ scripts/create-proposal.ts
@@ -11,12 +11,19 @@ import { assert } from "console";
 import { Arguments } from 'yargs';
 import { url } from "inspector";
 
+import { printDiscordProposal } from "./util";
+
 const yargs = require("yargs/yargs");
 const { hideBin } = require("yargs/helpers");
 const args = yargs(hideBin(process.argv))
     .option("url", {
         type: "string",
         description: "Polkadotjs url",
+        demandOption: true,
+    })
+    .option("description", {
+        type: "string",
+        description: "Description of the proposal",
         demandOption: true,
     })
     .argv;
@@ -36,33 +43,14 @@ async function main(): Promise<void> {
     console.log("extrinsic:", extrinsic);
 
     const api = await createSubstrateAPI(rpc as string);
+    
     const call = api.createType('Call', extrinsic);
     console.log('Full input tx: ', call.toHuman());
     console.log('');
     console.log('Abbreviated input tx: ', call.section.toString() + '.' + call.method.toString());
     console.log('');
 
-    const deposit = api.consts.democracy.minimumDeposit.toNumber();
-
-    let proposal;
-
-    if (call.toU8a().byteLength <= 128) {
-        proposal = api.tx.democracy.propose({ Inline: call.toHex() }, deposit);
-    } else {
-        const preImageSubmission = api.tx.preimage.notePreimage(call.toHex());
-        const innerProposal = api.tx.democracy.propose({ Lookup: {
-            hash: call.hash.toHex(),
-            len: call.toU8a().byteLength
-        }}, deposit);
-        proposal = api.tx.utility.batchAll([preImageSubmission, innerProposal]);
-    }
-
-    console.log('**Extrinsic:**', url.href);
-    console.log('');
-
-    url.hash = hashPrefix + proposal.toHex();
-    console.log('**Proposal:**', url.href);
-    console.log('');
+    printDiscordProposal(args["description"], api.tx(call), rpc as string, api);
 
     await api.disconnect();
 }
